jfkdls

+++ src/gui/dialogue/AntDialog.java
@@ -18,12 +18,12 @@ public class AntDialog extends Dialog {
 		try {
 			int width = Integer.parseInt(myTexts.get("width").getText().trim());
 			int height = Integer.parseInt(myTexts.get("height").getText().trim());
-			double percentAnts = Integer.parseInt(myTexts.get("percentAnts").getText().trim());
+			double percentAnts = Double.parseDouble(myTexts.get("percentAnts").getText().trim());
 			double evapRate = Double.parseDouble(myTexts.get("evapRate").getText().trim());
 			double diffusionRate = Double.parseDouble(myTexts.get("diffusionRate").getText().trim());
 
 			return width <= 100 && width > 0 && height <= 100 && height > 0 
-					&& percentAnts>=0 && percentAnts <= 1
+					&& percentAnts>= 0 && percentAnts <= 1
 					&& evapRate>=0 && evapRate <= 1
 					&& diffusionRate <= 1 && diffusionRate >= 0;
 		} catch (Exception e) {
